Clarified and improved language.
Eliminated reference to "unprofessional", as interpretations of the word vary too wildly. I don't like the "word" "obligated", preferring "obliged". Changed the overall language of the second paragraph to imply that the desired alternative to constructive criticism is silence.

Eliminated reference to "unacceptable" for the same reason as eliminating "unprofessional".