Redesign proposal
Tweaks to typography, layout, and color scheme.
I tried to make it easier to read, with a proportional font and larger size.
I also reduced the maximum line length a bit.
An SVG logo has been added. I'm not a graphic designer, but did my best.
Google fonts stylesheet and Google Analytics script moved to the bottom, to
prevent blocking.