Improved readability, en-GB version
As per
https://github.com/ContributorCovenant/contributor_covenant/issues/498

I've added a 1.4.2 version which contains an easier to read version.
I've included en-GB (confusingly .uk is Ukrainian,  gb is UK English)
and a US English version.

This version was originally written to comply with the UK Government
Digital Service Guidelines.  We think it is clearer and easier to
understand.