"open source" changed to "open" when describing CC license
Suggest that "open source" is incorrect for describing this project's Creative Commons documents which are text only. "Open Source" implies source code.
I propose that "free culture" be replaced by "free", as "free culture" is not defined within this document and will be confusing to many readers, especially readers who are new to Open Source communities.