Change disability term by functional diversity
This commit changes the term ‘disability’ by ‘functional diversity’. Functional diversity was proposed as a new term in the struggle for dignity in the diversity of the human being on the Independent Living Forum (Spain) - May 2005.

It has been only changed for en and es languages, but it must be changed in other languages as well.