Update home page block linking to v2.0
This commit will change the links in the home page for accessing to the
latest version, so that they point to 2.0 rather than to 1.4.

As a companion to commit fc5fd23f2576dbe0cd75c6b0dfc91d4c720513b6
(Add 1.4 to the list of previous versions), this should make all the
published versions of the code of conduct available through the home
page.