Ask for tolerance, not respect
Respect cannot be manufactured at will. If you don't respect an idea
(for example that the Earth is flat), then it doesn't matter how hard
you try; you still will not respect it. In that sense respect is like
belief; nobody can force you to believe the Moon is made of cheese.
Either you do or you don't.

You can pretend to believe in something, and you can pretend to respect
something, but you really don't. Any policy that asks people to pretend
is not a good policy.

The commitment should be towards tolerance; which can be honestly given,
not respect; which can only be faked.

A nuanced debate over tolerance versus respect can be seen in the
proposal at Cambridge University to demand respect. Tolerance won.

Fixes #892.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>